erofs-utils: lib: introduce liberofs_global_{init,exit}
Currently, the codebase only supports a single instance during
the entire running time. This will be extended to allow 3rd-party
applications to formally use liberofs with multiple instances
simultaneously. Therefore, it is necessary to separate global
initialization from per-instance initialization.

Another point is that libcurl will be used for other remote
backends, so it also needs to be separated from the current S3
backend.

Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Link: https://lore.kernel.org/r/20250819032818.3598157-1-hsiangkao@linux.alibaba.com