erofs-utils: mount: add NBD backend support
Currently, only local images are supported (so using the current NBD
backend is mostly for functional testing), e.g.:

 $ sudo mount.erofs -t erofs.nbd foo.erofs mnt/

However, it already uses the vfile interface, so extending it to
remote sources should be straightforward.

NBD failover is not supported yet and should be implemented via the
new netlink interface.

Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Link: https://lore.kernel.org/r/20250827061139.1310-1-hsiangkao@linux.alibaba.com